+++ src/base/combination/combo.js
@@ -9,6 +9,9 @@
             var conf = BI.Combo.superclass._defaultConfig.apply(this, arguments);
             return BI.extend(conf, {
                 baseCls: (conf.baseCls || "") + " bi-combo" + (BI.isIE() ? " hack" : ""),
+                attributes: {
+                    tabIndex: -1
+                },
                 trigger: "click",
                 toggle: true,
                 direction: "bottom", // top||bottom||left||right||top,left||top,right||bottom,left||bottom,right||right,innerRight||right,innerLeft||innerRight||innerLeft
@@ -214,6 +217,43 @@
                                     }
                                 }, 50);
                             }
+                        });
+                        break;
+                    case "click-blur":
+                        var debounce = BI.debounce(function (e) {
+                            if (self.combo.element.__isMouseInBounds__(e)) {
+                                if (self.isEnabled() && self.isValid() && self.combo.isEnabled() && self.combo.isValid()) {
+                                    // if (!o.toggle && self.isViewVisible()) {
+                                    //     return;
+                                    // }
+                                    o.toggle ? self._toggle(e) : self._popupView(e);
+                                    if (self.isViewVisible()) {
+                                        self.fireEvent(BI.Controller.EVENT_CHANGE, BI.Events.EXPAND, "", self.combo);
+                                        self.fireEvent(BI.Combo.EVENT_EXPAND);
+                                    } else {
+                                        self.fireEvent(BI.Controller.EVENT_CHANGE, BI.Events.COLLAPSE, "", self.combo);
+                                        self.fireEvent(BI.Combo.EVENT_COLLAPSE);
+                                    }
+                                }
+                            }
+                        }, BI.EVENT_RESPONSE_TIME, {
+                            "leading": true,
+                            "trailing": false
+                        });
+                        self.element.off("click." + self.getName()).on("click." + self.getName(), function (e) {
+                            debounce(e);
+                            try {
+                                self.element[0].focus();
+                            } catch (e) {
+
+                            }
+                            st(e);
+                        });
+                        self.element.off("blur." + self.getName()).on("blur." + self.getName(), function (e) {
+                            if (self.isViewVisible()) {
+                                self._hideView(e);
+                            }
+                            st(e);
                         });
                         break;
                 }
@@ -528,11 +568,14 @@
         },
 
         destroyed: function () {
-            BI.Widget._renderEngine.createElement(document).unbind("mousedown." + this.getName())
+            BI.Widget._renderEngine.createElement(document)
+                .unbind("click." + this.getName())
+                .unbind("mousedown." + this.getName())
                 .unbind("mousewheel." + this.getName())
                 .unbind("mouseenter." + this.getName())
                 .unbind("mousemove." + this.getName())
-                .unbind("mouseleave." + this.getName());
+                .unbind("mouseleave." + this.getName())
+                .unbind("blur." + this.getName());
             BI.Resizers.remove(this.getName());
             this.popupView && this.popupView._destroy();
             delete needHideWhenAnotherComboOpen[this.getName()];