+++ src/component/allvaluechooser/abstract.allvaluechooser.js
@@ -82,5 +82,20 @@ BI.AbstractAllValueChooser = BI.inherit(BI.Widget, {
                 hasNext: false
             });
         }
-    }
+    },
+
+    _assertValue: function (v) {
+        v = v || {};
+        var value = v;
+        if (BI.isNotNull(this.items)) {
+            var isAllSelect = BI.difference(BI.map(this.items, "value"), v.value).length === 0;
+            if (isAllSelect) {
+                value = {
+                    type: BI.Selection.All,
+                    value: [],
+                };
+            }
+        }
+        return value;
+    },
 });