+++ src/core/wrapper/layout/layout.card.js
@@ -28,7 +28,7 @@ BI.CardLayout = BI.inherit(BI.Layout, {
         BI.each(items, function (i, item) {
             if (item) {
                 if (!self.hasWidget(item.cardName)) {
-                    var w = BI.createWidget(item);
+                    var w = BI._lazyCreateWidget(item);
                     w.on(BI.Events.DESTROY, function () {
                         var index = BI.findIndex(o.items, function (i, tItem) {
                             return tItem.cardName == item.cardName;
@@ -98,7 +98,7 @@ BI.CardLayout = BI.inherit(BI.Layout, {
         if (this.isCardExisted(cardName)) {
             throw new Error("cardName is already exist");
         }
-        var widget = BI.createWidget(cardItem, this);
+        var widget = BI._lazyCreateWidget(cardItem, this);
         widget.element.css({
             position: "relative",
             top: "0",